Fix malformed material element
The URDF specification allows to have a material element without a color child element,
but only if the name of the color is specified in the model-wide material database.
Otherwise, it is always necessary to specify the color of the material of the link.

Fix https://github.com/robotology/icub-model-generator/issues/36
See https://github.com/robotology/icub-model-generator/issues/36#issuecomment-313065203